NP-988 Use same component for all listing of registrations
